出品者本人でないと削除できないよう条件を設定

+++ app/controllers/products_controller.rb
@@ -24,8 +24,10 @@ class ProductsController < ApplicationController
 
   def destroy
     product = Product.find(params[:id])
-    product.destroy
-    redirect_to root_path
+    if current_user.id == product.user_id
+      product.destroy
+      redirect_to root_path
+    end
   end
 
   private
